improve formatting
