+++ src/helpers.ts
@@ -1,0 +1,3 @@
+export const asQuote = (str: string) => {
+  return str.split('\n').map((line) => `> ${line}`).join('\n')
+}
